// src/utils/regexp/pattern-type.ts
export const enum PatternType {
  Backreference = 'Backreference',
  BoundaryAssertion = 'BoundaryAssertion',
  CapturingGroup = 'CapturingGroup',
  CharacterSequence = 'CharacterSequence',
  Concatenation = 'Concatenation',
  Disjunction = 'Disjunction',
  LookaroundAssertion = 'LookaroundAssertion',
  NonCapturingGroup = 'NonCapturingGroup',
  Quantifier = 'Quantifier',
}